pyproject.toml
[tool.poetry]
name = "projects-manager"
version = "0.1.0"
description = "Api to manage plot of land."
authors = ["Sebastian Wdowiarz <wdowiarzsebastian@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
geoalchemy2 = "^0.16.0"
geojson = "^3.2.0"
asyncpg = "^0.30.0"
shapely = "^2.0.6"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.14.1"
isort = "^5.13.2"
pre-commit = "^4.0.1"
flake8 = "^7.1.1"
bandit = "^1.8.0"
pyupgrade = "^3.19.1"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
target-version = ['py311']
